+++ src/steps/fetch-administrator-roles/__tests__/index.test.ts
@@ -1,144 +1,23 @@
 import {
+  executeStepWithDependencies,
   Recording,
-  setupRecording,
-  createMockStepExecutionContext,
-} from '../../../../test';
+} from '@jupiterone/integration-sdk-testing';
+import { buildStepTestConfigForStep } from '../../../../test/config';
+import { setupProjectRecording } from '../../../../test/recording';
 
-import step, { createAdministratorRoleEntity } from '../index';
-import {
-  createDeepSecurityClient,
-  DeepSecurityAdministratorRole,
-} from '../../../provider';
-import { TrendMicroIntegrationConfig } from '../../../types';
-
+// See test/README.md for details
 let recording: Recording;
-
-beforeEach(() => {
-  recording = setupRecording({
-    directory: __dirname,
-    name: 'administratorRoles',
-  });
-});
-
 afterEach(async () => {
   await recording.stop();
 });
 
-test('administratorRole fetching', async () => {
-  const context = createMockStepExecutionContext<TrendMicroIntegrationConfig>({
-    instanceConfig: {
-      apiKey: 'apiKey',
-    },
+test('fetch-administrator-roles', async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: 'fetch-administrator-roles',
   });
-  const provider = createDeepSecurityClient(context.instance);
 
-  const results = await provider.listAdministratorRoles();
-
-  expect(results).toEqual({
-    roles: expect.arrayContaining([
-      expect.objectContaining({
-        name: 'Full Access',
-        description: '',
-        urn: 'urn:tmds:identity:us-east-ds-1:78422:role/Full Access',
-        immutable: true,
-        canOnlyManipulateUsersWithEqualOrLesserRights: false,
-        allComputers: true,
-        allPolicies: true,
-        allowUserInterface: true,
-        allowWebService: true,
-        ID: 1,
-      }),
-      expect.objectContaining({
-        name: 'Auditor',
-        description: '',
-        urn: 'urn:tmds:identity:us-east-ds-1:78422:role/Auditor',
-        immutable: false,
-        canOnlyManipulateUsersWithEqualOrLesserRights: false,
-        allComputers: true,
-        allPolicies: true,
-        allowUserInterface: true,
-        allowWebService: true,
-        ID: 2,
-      }),
-      expect.objectContaining({
-        name: 'New Role',
-        description: '',
-        urn: 'urn:tmds:identity:us-east-ds-1:78422:role/New Role',
-        immutable: false,
-        canOnlyManipulateUsersWithEqualOrLesserRights: false,
-        allComputers: true,
-        allPolicies: true,
-        allowUserInterface: true,
-        allowWebService: false,
-        ID: 3,
-      }),
-      expect.objectContaining({
-        name: 'New Role_2',
-        description: '',
-        urn: 'urn:tmds:identity:us-east-ds-1:78422:role/New Role_2',
-        immutable: false,
-        canOnlyManipulateUsersWithEqualOrLesserRights: false,
-        allComputers: true,
-        allPolicies: true,
-        allowUserInterface: true,
-        allowWebService: false,
-        ID: 4,
-      }),
-    ]),
-  });
-});
-
-test('administator role entity conversion', () => {
-  const role = {
-    ID: 4,
-    name: 'New Role_2',
-    description: '',
-    urn: 'urn:tmds:identity:us-east-ds-1:78422:role/New Role_2',
-    immutable: false,
-    canOnlyManipulateUsersWithEqualOrLesserRights: false,
-    allComputers: true,
-    allPolicies: true,
-    allowUserInterface: true,
-    allowWebService: false,
-  } as DeepSecurityAdministratorRole;
-
-  expect(createAdministratorRoleEntity(role)).toEqual({
-    name: 'New Role_2',
-    id: 'trend-micro-administrator-role:4',
-    description: '',
-    urn: 'urn:tmds:identity:us-east-ds-1:78422:role/New Role_2',
-    _key: 'trend-micro-administrator-role:4',
-    _type: 'trend_micro_administrator_role',
-    _class: ['AccessRole'],
-    createdOn: undefined,
-    displayName: 'New Role_2',
-    _rawData: [{ name: 'default', rawData: expect.objectContaining(role) }],
-  });
-});
-
-test('step data collection', async () => {
-  const context = createMockStepExecutionContext<TrendMicroIntegrationConfig>({
-    instanceConfig: {
-      apiKey: 'apiKey',
-    },
-  });
-  await step.executionHandler(context);
-
-  expect(context.jobState.collectedEntities).toHaveLength(4);
-  expect(context.jobState.collectedRelationships).toHaveLength(0);
-
-  expect(context.jobState.collectedEntities).toEqual([
-    expect.objectContaining({
-      _key: 'trend-micro-administrator-role:1',
-    }),
-    expect.objectContaining({
-      _key: 'trend-micro-administrator-role:2',
-    }),
-    expect.objectContaining({
-      _key: 'trend-micro-administrator-role:3',
-    }),
-    expect.objectContaining({
-      _key: 'trend-micro-administrator-role:4',
-    }),
-  ]);
+  const stepConfig = buildStepTestConfigForStep('fetch-administrator-roles');
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
 });