+++ src/steps/fetch-administrators/__tests__/index.test.ts
@@ -1,168 +1,23 @@
 import {
+  executeStepWithDependencies,
   Recording,
-  setupRecording,
-  createMockStepExecutionContext,
-} from '../../../../test';
+} from '@jupiterone/integration-sdk-testing';
+import { buildStepTestConfigForStep } from '../../../../test/config';
+import { setupProjectRecording } from '../../../../test/recording';
 
-import step, { createAdministratorEntity } from '../index';
-import {
-  createDeepSecurityClient,
-  DeepSecurityAdministrator,
-} from '../../../provider';
-import { TrendMicroIntegrationConfig } from '../../../types';
-
+// See test/README.md for details
 let recording: Recording;
-
-beforeEach(() => {
-  recording = setupRecording({
-    directory: __dirname,
-    name: 'administrators',
-  });
-});
-
 afterEach(async () => {
   await recording.stop();
 });
 
-test('administrator fetching', async () => {
-  const context = createMockStepExecutionContext<TrendMicroIntegrationConfig>({
-    instanceConfig: {
-      apiKey: 'apiKey',
-    },
+test('fetch-administrators', async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: 'fetch-administrators',
   });
-  const provider = createDeepSecurityClient(context.instance);
 
-  const results = await provider.listAdministrators();
-
-  expect(results).toEqual({
-    administrators: expect.arrayContaining([
-      expect.objectContaining({
-        username: 'charlie.duong@jupiterone.io',
-        fullName: '',
-        description: '',
-        roleID: 1,
-        locale: 'en-US',
-        timeZone: 'US/Eastern',
-        timeFormat: '24',
-        passwordNeverExpires: true,
-        active: true,
-        mfaType: 'none',
-        phoneNumber: '',
-        mobileNumber: '',
-        pagerNumber: '',
-        emailAddress: 'charlie.duong@jupiterone.io',
-        primaryContact: true,
-        receiveNotifications: false,
-        reportPDFPasswordEnabled: false,
-        created: 1586982455921,
-        lastPasswordChange: 1586982455921,
-        lastSignIn: 1587488991381,
-        unsuccessfulSignInAttempts: 0,
-        external: false,
-        type: 'normal',
-        readOnly: false,
-        ID: 1,
-        UTCOffset: 'UTC-4.00, DST',
-      }),
-      expect.objectContaining({
-        username: 'charlie.duong.test',
-        fullName: 'Charlie Test',
-        description: 'Testing123',
-        roleID: 2,
-        locale: 'en-US',
-        timeZone: 'US/Eastern',
-        timeFormat: '24',
-        passwordNeverExpires: false,
-        active: true,
-        mfaType: 'none',
-        phoneNumber: '',
-        mobileNumber: '',
-        pagerNumber: '',
-        emailAddress: '',
-        primaryContact: false,
-        receiveNotifications: false,
-        reportPDFPasswordEnabled: false,
-        created: 1587489647292,
-        lastPasswordChange: 1587489647292,
-        lastSignIn: 1587489661720,
-        unsuccessfulSignInAttempts: 0,
-        external: false,
-        type: 'normal',
-        readOnly: false,
-        ID: 8,
-        UTCOffset: 'UTC-4.00, DST',
-      }),
-    ]),
-  });
-});
-
-test('administator entity conversion', () => {
-  const admin = {
-    username: 'charlie.duong.test',
-    fullName: 'Charlie Test',
-    description: 'Testing123',
-    roleID: 2,
-    locale: 'en-US',
-    timeZone: 'US/Eastern',
-    timeFormat: '24',
-    passwordNeverExpires: false,
-    active: true,
-    mfaType: 'none',
-    phoneNumber: '',
-    mobileNumber: '',
-    pagerNumber: '',
-    emailAddress: '',
-    primaryContact: false,
-    receiveNotifications: false,
-    reportPDFPasswordEnabled: false,
-    created: 1587489647292,
-    lastPasswordChange: 1587489647292,
-    lastSignIn: 1587489661720,
-    unsuccessfulSignInAttempts: 0,
-    external: false,
-    type: 'normal',
-    readOnly: false,
-    ID: 8,
-    UTCOffset: 'UTC-4.00, DST',
-  } as DeepSecurityAdministrator;
-
-  expect(createAdministratorEntity(admin)).toEqual({
-    description: 'Testing123',
-    active: true,
-    _key: 'trend-micro-administrator:8',
-    _type: 'trend_micro_administrator',
-    _class: ['User'],
-    name: 'Charlie Test',
-    createdOn: 1587489647292,
-    username: 'charlie.duong.test',
-    roleId: 2,
-    _rawData: [
-      {
-        name: 'default',
-        rawData: admin,
-      },
-    ],
-    displayName: 'Charlie Test',
-  });
-});
-
-test('step data collection', async () => {
-  const context = createMockStepExecutionContext<TrendMicroIntegrationConfig>({
-    instanceConfig: {
-      apiKey: 'apiKey',
-    },
-  });
-  await step.executionHandler(context);
-
-  expect(context.jobState.collectedEntities).toHaveLength(2);
-  expect(context.jobState.collectedRelationships).toHaveLength(0);
-
-  expect(context.jobState.collectedEntities).toEqual([
-    expect.objectContaining({
-      _key: 'trend-micro-administrator:1',
-    }),
-    expect.objectContaining({
-      _key: 'trend-micro-administrator:8',
-    }),
-  ]);
+  const stepConfig = buildStepTestConfigForStep('fetch-administrators');
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
 });